refactor: move main and packages to cmd and internal
