+++ internal/utils/binary.go
